+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2026-05-17
+
 ### Added
 
 - Member suspension feature
@@ -367,7 +369,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    - Push to remote
    - Trigger automated deployment
 
-[Unreleased]: https://github.com/jnahian/daily-dose/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/jnahian/daily-dose/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/jnahian/daily-dose/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/jnahian/daily-dose/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/jnahian/daily-dose/compare/v1.4.6...v1.5.0
 [1.4.6]: https://github.com/jnahian/daily-dose/compare/v1.4.5...v1.4.6